+++ src/block.rs
@@ -1,4 +1,5 @@
 use super::*;
+use crate::errors::Result;
 use bincode::{deserialize, serialize};
 use colored::*;
 use crypto::digest::Digest;
@@ -6,8 +7,7 @@ use crypto::sha2::Sha256;
 use log::{debug, info};
 use std::time::SystemTime;
 
-const TARGET_HEX: usize = 4;
-pub type Result<T> = std::result::Result<T, failure::Error>;
+pub const TARGET_HEX: usize = 4;
 
 #[derive(Debug, Clone)]
 pub struct Block {
@@ -17,11 +17,6 @@ pub struct Block {
     hash: String,
     prev_block_hash: String,
     transactions: String,
-}
-
-#[derive(Debug, Clone)]
-pub struct Blockchain {
-    blocks: Vec<Block>,
 }
 
 impl Block {
@@ -43,7 +38,7 @@ impl Block {
         Ok(block)
     }
 
-    fn get_hash(&self) -> String {
+    pub fn get_hash(&self) -> String {
         self.hash.clone()
     }
 
@@ -72,7 +67,7 @@ impl Block {
         Ok(bytes)
     }
 
-    fn new_genesis_block() -> Block {
+    pub fn new_genesis_block() -> Block {
         Block::new_block(String::from("Genesis Block"), String::new(), 0).unwrap()
     }
 
@@ -85,36 +80,5 @@ impl Block {
         println!("{}", format!("{:#?}", vec1).green());
 
         Ok(&hasher.result_str()[0..TARGET_HEX] == String::from_utf8(vec1)?)
-    }
-}
-
-impl Blockchain {
-    pub fn new() -> Blockchain {
-        Blockchain {
-            blocks: vec![Block::new_genesis_block()],
-        }
-    }
-
-    pub fn add_block(&mut self, data: String) -> Result<()> {
-        let prev_block = self.blocks.last().unwrap();
-        let new_block = Block::new_block(data, prev_block.get_hash(), TARGET_HEX)?;
-        self.blocks.push(new_block);
-        Ok(())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_blockchain() {
-        let mut b = Blockchain::new();
-        b.add_block("data".to_string());
-        b.add_block("data2".to_string());
-        b.add_block("data3".to_string());
-        println!("{}", format!("{:#?}", b).green());
-
-        dbg!(b);
     }
 }
