+++ src/main.rs
@@ -1,5 +1,6 @@
 pub mod block;
-
+pub mod blockchain;
+pub mod errors;
 fn main() {
     println!("Hello, world!");
     let u: Vec<i32> = vec![];
